requirements.txt
pypdf
tdqm
openai>=0.27
